@@ -1,12 +1,12 @@
 package com.dscvit.android.devfest18.di
 
 import android.app.Activity
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.app.Application
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentActivity
-import android.support.v4.app.FragmentManager
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.fragment.app.FragmentManager
 import com.dscvit.android.devfest18.DevfestApp
 import dagger.android.AndroidInjection
 import dagger.android.support.AndroidSupportInjection
