@@ -1,7 +1,7 @@
 package com.dscvit.android.devfest18.di.module
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import com.dscvit.android.devfest18.di.key.ViewModelKey
 import com.dscvit.android.devfest18.ui.FactoryViewModel
 import com.dscvit.android.devfest18.ui.main.MainViewModel
